Move BasePartitionerSharedState into separate header
